Notes for updating
******************

Linux:
You don't need an extra package for CEGUILua and Tolua, it's already shipped with CEGUI.
However you do need to make sure that the OgreRenderer is installed too with CEGUI 0.7 (may be a separate package).
Also, Orxonox now recognises if you install the CgProgramManager (a separate package available on newer Ubuntu on Debian systems).

Windows:
Download the new dependency packages versioned 6.0 and use these. If you have problems with that or if you don't like the in game console problem mentioned below, you can download the new 4.3 version of the packages (only available for Visual Studio 2005/2008).


Key new features:
- ***Support for Mac OS X***
- Visual Studio 2010 support
- Bullet library update to 2.77
- OIS library update to 1.3
- Support for CEGUI 0.7
  --> Support for Arch Linux and even SuSE
- Improved install target
- Compiles now with GCC 4.6
- Ogre Cg Shader plugin activated for Linux if available
- And of course lots of bug fixes

There are also some regressions:
- No support for CEGUI 0.5, Ogre 1.4 and boost 1.35 - 1.39 any more
- In game console is not working in main menu for CEGUI 0.7
- Tolua (just the C lib, not the application) and CEGUILua libraries are no longer in our repository.
  --> You will need to get these as well when compiling Orxonox
- And of course lots of new bugs we don't yet know about

+++ code/trunk/src/libraries/core/input/KeyBinder.cc
@@ -395,9 +395,9 @@ namespace orxonox
             Button* button = it->second;
 
             // iterate over all modes
-            for (size_t mode_index = 0; mode_index < 3; ++mode_index)
+            for (int mode_index = 0; mode_index < 3; ++mode_index)
             {
-                if (mode_index == (unsigned int)new_mode) // skip commands that are already in the desired mode
+                if (mode_index == new_mode) // skip commands that are already in the desired mode
                     continue;
 
                 // iterate over all commands of the given mode at the given button
